--- app/fabric.properties
@@ -1,3 +1,0 @@
-#Contains API Secret used to validate your application. Commit to internal source control; avoid making secret public.
-#Thu Jun 23 19:30:29 EEST 2016
-apiSecret=2d97f10272b6f88c01d479e31b2b18182ee2aced7cc9970958215160adb1324e
