+++ app/src/main/java/com/larin92/testtasks/internship/api/ApiSettings.java
@@ -1,0 +1,15 @@
+package com.larin92.testtasks.internship.api;
+
+/**
+ * Class for constants, used for URL completing for REST requests
+ */
+public class ApiSettings {
+
+    private static final String SCHEME = "http://";
+
+    private static final String HOSTNAME = "dev-contact.yalantis.com/";
+
+    private final static String ENDPOINT = "rest/v1/";
+
+    public static final String SERVER = SCHEME + HOSTNAME + ENDPOINT;
+}
