+++ app/src/main/java/com/larin92/testtasks/internship/model/CardModel.java
@@ -1,4 +1,4 @@
-package com.larin92.testtasks.internship.data.model;
+package com.larin92.testtasks.internship.model;
 
 import java.util.ArrayList;
