+++ app/src/main/java/com/larin92/testtasks/internship/model/ProfileFB.java
@@ -1,4 +1,4 @@
-package com.larin92.testtasks.internship.data.model;
+package com.larin92.testtasks.internship.model;
 
 import android.net.Uri;
