+++ app/src/main/java/com/larin92/testtasks/internship/model/json/City.java
@@ -1,4 +1,4 @@
-package com.larin92.testtasks.internship.data.model.json;
+package com.larin92.testtasks.internship.model.json;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
