+++ app/src/main/java/com/larin92/testtasks/internship/presenter/ItemPresenter.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 
 import com.larin92.testtasks.internship.App;
 import com.larin92.testtasks.internship.contract.ItemContract;
-import com.larin92.testtasks.internship.data.model.CardModel;
+import com.larin92.testtasks.internship.model.CardModel;
 
 public class ItemPresenter implements ItemContract.Presenter {
