+++ app/src/main/java/com/larin92/testtasks/internship/ui/adapters/CardsRecyclerAdapter.java
@@ -7,7 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.larin92.testtasks.internship.R;
-import com.larin92.testtasks.internship.data.model.CardModel;
+import com.larin92.testtasks.internship.model.CardModel;
 import com.larin92.testtasks.internship.ui.fragments.NavigationFragment;
 
 import java.util.List;
