+++ app/src/main/java/com/larin92/testtasks/internship/ui/adapters/Holder.java
@@ -8,7 +8,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.larin92.testtasks.internship.R;
-import com.larin92.testtasks.internship.data.model.CardModel;
+import com.larin92.testtasks.internship.model.CardModel;
 import com.larin92.testtasks.internship.ui.fragments.NavigationFragment;
 
 import butterknife.BindView;
