+++ app/src/main/java/com/larin92/testtasks/internship/ui/fragments/NavigationFragment.java
@@ -9,10 +9,11 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import com.larin92.testtasks.internship.R;
 import com.larin92.testtasks.internship.contract.NavigationContract;
-import com.larin92.testtasks.internship.data.model.CardModel;
+import com.larin92.testtasks.internship.model.CardModel;
 import com.larin92.testtasks.internship.presenter.NavigationPresenter;
 import com.larin92.testtasks.internship.ui.adapters.CardsRecyclerAdapter;
 import com.melnykov.fab.FloatingActionButton;
@@ -132,6 +133,7 @@ public class NavigationFragment extends Fragment implements NavigationContract.V
 
     @Override
     public void notifyAdapter(List<CardModel> list) {
+        Toast.makeText(getActivity(), "i'm notifying", Toast.LENGTH_SHORT).show();
         mRecyclerAdapter.addCards(list);
         mRecyclerAdapter.notifyDataSetChanged();
     }
@@ -206,6 +208,7 @@ public class NavigationFragment extends Fragment implements NavigationContract.V
         mNavigationPresenter.detachView();
         mUnbinder.unbind();
         mRecyclerAdapter = null;
+        mLinearLayoutManager = null;
     }
 
     /**
